requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.4.1
idna==2.10
multidict==4.7.6
requests==2.24.0
urllib3==1.25.10
yarl==1.6.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.4.1
idna==2.10
multidict==4.7.6
psycopg2-binary==2.8.6
requests==2.24.0
urllib3==1.25.10
yarl==1.6.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
discord.py==1.4.1
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.2
multidict==4.7.6
numpy==1.19.2
Pillow==7.2.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
yarl==1.6.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
discord.py==1.4.1
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.2
multidict==4.7.6
numpy==1.19.2
Pillow==7.2.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
yarl==1.6.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
discord.py==1.4.1
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.2
multidict==4.7.6
numpy==1.19.2
Pillow==7.2.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
yarl==1.6.0
